+++ src/main/java/wooteco/subway/member/dao/MemberDao.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
 import wooteco.subway.member.domain.Member;
 
 import javax.sql.DataSource;
+import java.util.Optional;
 
 @Repository
 public class MemberDao {
@@ -57,8 +58,10 @@ public class MemberDao {
         return jdbcTemplate.queryForObject(sql, Integer.class, email);
     }
 
-    public Member findByEmail(String email) {
+    public Optional<Member> findByEmail(String email) {
         String sql = "select * from MEMBER where email = ?";
-        return jdbcTemplate.queryForObject(sql, rowMapper, email);
+        return jdbcTemplate.query(sql, rowMapper, email)
+                .stream()
+                .findAny();
     }
 }
