+++ src/main/java/wooteco/subway/member/domain/Member.java
@@ -1,7 +1,7 @@
 package wooteco.subway.member.domain;
 
 import org.apache.commons.lang3.StringUtils;
-import wooteco.subway.auth.application.AuthorizationException;
+import wooteco.subway.exception.auth.WrongPasswordException;
 
 public class Member {
     private Long id;
@@ -49,7 +49,7 @@ public class Member {
 
     public void checkPassword(String password) {
         if (!StringUtils.equals(this.password, password)) {
-            throw new AuthorizationException();
+            throw new WrongPasswordException();
         }
     }
 }
